proj 실행 시 data 생성 ã user, question
